python-client/requirements.txt
grpcio
grpcio-tools
pika
prometheus_client
